Move all subcommands out of main into proper folders
