+++ client/rpc/root.go
@@ -1,10 +1,16 @@
-package main
+package rpc
 
 import (
+	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 
 	"github.com/cosmos/cosmos-sdk/client"
 )
+
+// XXX: remove this when not needed
+func todoNotImplemented(_ *cobra.Command, _ []string) error {
+	return errors.New("TODO: Command not yet implemented")
+}
 
 const (
 	// one of the following should be provided to verify the connection
@@ -13,11 +19,6 @@ const (
 	flagValHash = "validator-set"
 
 	flagSelect = "select"
-	flagTags   = "tag"
-	flagAny    = "any"
-
-	flagBind = "bind"
-	flagCORS = "cors"
 )
 
 var (
@@ -28,19 +29,13 @@ var (
 	}
 )
 
-// AddClientCommands returns a sub-tree of all basic client commands
-//
-// Call AddGetCommand and AddPostCommand to add custom txs and queries
-func AddClientCommands(cmd *cobra.Command) {
+// AddCommands adds a number of rpc-related subcommands
+func AddCommands(cmd *cobra.Command) {
 	cmd.AddCommand(
 		initClientCommand(),
 		statusCmd,
 		blockCommand(),
 		validatorCommand(),
-		lineBreak,
-		txSearchCommand(),
-		txCommand(),
-		lineBreak,
 	)
 }
 
@@ -72,38 +67,6 @@ func validatorCommand() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "validatorset <height>",
 		Short: "Get the full validator set at given height",
-		RunE:  todoNotImplemented,
-	}
-	return cmd
-}
-
-func serveCommand() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "serve",
-		Short: "Start LCD (light-client daemon), a local REST server",
-		RunE:  todoNotImplemented,
-	}
-	// TODO: handle unix sockets also?
-	cmd.Flags().StringP(flagBind, "b", "localhost:1317", "Interface and port that server binds to")
-	cmd.Flags().String(flagCORS, "", "Set to domains that can make CORS requests (* for all)")
-	return cmd
-}
-
-func txSearchCommand() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "txs",
-		Short: "Search for all transactions that match the given tags",
-		RunE:  todoNotImplemented,
-	}
-	cmd.Flags().StringSlice(flagTags, nil, "Tags that must match (may provide multiple)")
-	cmd.Flags().Bool(flagAny, false, "Return transactions that match ANY tag, rather than ALL")
-	return cmd
-}
-
-func txCommand() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "tx <hash>",
-		Short: "Matches this txhash over all committed blocks",
 		RunE:  todoNotImplemented,
 	}
 	return cmd
